[wr8_ros] sync replaced, updated ros_lib data

+++ wr8_msgs/generate_uc_types.sh
@@ -7,6 +7,14 @@ if [ $ret -ne 0 ]; then
 	exit 1
 fi
 
+CALL_DIR=`dirname $0`
+if [ "$CALL_DIR" != "./wr8_msgs" ]; then
+	echo "Must be called inside base repository"
+	exit 1
+fi
+
 LIB_DST_DIR=`rospack find wr8_msgs`
 rm -rf $LIB_DST_DIR/ros_lib
-rosrun wr8_msgs make_uc_lib.py $LIB_DST_DIR $CATKIN_WS
+
+rosrun wr8_msgs make_uc_lib.py $LIB_DST_DIR $CATKIN_WS &&
+rsync -avzPc $LIB_DST_DIR/ros_lib/ controller_wr_driver/ros_lib/
